% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goview.R
\name{goview}
\alias{goview}
\title{goview Function}
\usage{
goview(goresults)
}
\arguments{
\item{goresults}{the go results from gProfileR}
}
\description{
This function allows you to show the results of GO enrichment analysis.
}
\examples{
goview()
}
\author{
yingying
}
\keyword{GO}
